refactor(ResultBoard): move to 'winning' directory for better modularity
