+++ src/main/java/lotto/SequenceManager.java
@@ -10,6 +10,7 @@ import lotto.winning.Analyst;
 import lotto.winning.BonusNumber;
 import lotto.winning.Prize;
 import lotto.winning.PrizeRankChecker;
+import lotto.winning.ResultBoard;
 import lotto.winning.Tally;
 import lotto.winning.WinningNumbers;
